backend/doctor/requirements.txt
fastapi
uvicorn
fastapi-cors
elasticsearch==8.16.0
python-dotenv
colorama